refactor : régorganisation des classes et fichiers
