+++ front/app/community/[siren]/comparison/[comparedSiren]/components/Header.tsx
