+++ front/app/community/[siren]/comparison/[comparedSiren]/page.tsx
@@ -2,11 +2,17 @@ import { Suspense } from 'react';
 
 import type { Metadata } from 'next';
 
+import { TransparencyScoreBar } from '@/components/TransparencyScore/TransparencyScore';
+import { TransparencyScore } from '@/components/TransparencyScore/constants';
 import Loading from '@/components/ui/Loading';
 import { fetchCommunities } from '@/utils/fetchers/communities/fetchCommunities-server';
+import { formatNumberInteger, stringifyCommunityType } from '@/utils/utils';
 
-import { ComparingFiche } from './ComparingFiche/ComparingFiche';
-import { Header } from './Header/Header';
+import { GlobalStatsComparison } from './components/GlobalStatsComparison';
+import { Header } from './components/Header';
+import { HeaderComparison } from './components/HeaderComparison';
+import { MarchesPublicsComparison } from './components/MarchesPublicsComparison';
+import { TransparenceComparison } from './components/TransparenceComparison';
 
 type PageProps = { params: Promise<{ siren: string; comparedSiren: string }> };
 
@@ -36,14 +42,17 @@ export default async function Page({ params }: PageProps) {
   const siren = (await params).siren;
   const siren2 = (await params).comparedSiren;
 
-  const community = await getCommunity(siren);
+  const community1 = await getCommunity(siren);
   const community2 = await getCommunity(siren2);
 
   return (
-    <Suspense key={community.siren + community2.siren} fallback={<Loading />}>
-      <Header community={community} community2={community2} />
-      <div className='mx-5 my-3 max-w-screen-xl mx-auto'>
-        <ComparingFiche community1={community} community2={community2} />
+    <Suspense key={community1.siren + community2.siren} fallback={<Loading />}>
+      <Header community={community1} community2={community2} />
+      <div className='mx-5 mx-auto my-3 max-w-screen-xl'>
+        <HeaderComparison community1={community1} community2={community2} />
+        <GlobalStatsComparison community1={community1} community2={community2} />
+        <TransparenceComparison community1={community1} community2={community2} />
+        <MarchesPublicsComparison siren1={community1.siren} siren2={community2.siren} />
       </div>
     </Suspense>
   );
